Test 'invalid extend'
