+++ generator/test/generator_test_io/invalid_extend.dart
@@ -1,0 +1,11 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'invalid_extend.g.dart';
+
+class Base {}
+
+@RealmModel()
+class _Bad extends Base {
+  @PrimaryKey()
+  late int id;
+}
