+++ generator/test/generator_test_io/invalid_extend.log
@@ -1,0 +1,8 @@
+Realm model classes can only extend Object
+
+in: asset:pkg/test/generator_test_io/invalid_extend.dart:8:7
+  ╷
+7 │ @RealmModel()
+8 │ class _Bad extends Base {
+  │       ^^^^ cannot extend Base
+  ╵
